+++ packages/dropdown/src/index.test.tsx
@@ -100,3 +100,59 @@ describe('Dropdown onChange', () => {
     expect(onChange.mock.calls).toEqual([['2']]);
   });
 });
+
+describe('Dropdown onFocus', () => {
+  const onFocus = jest.fn();
+
+  const wrapper = mount(
+    <Theme>
+      <Dropdown size="md" items={[{ value: '1', label: '1' }, { value: '2', label: '2' }]} onFocus={onFocus} />
+    </Theme>
+  );
+
+  it('should call onFocus prop', () => {
+    const select = wrapper.find(Select);
+    select.simulate('focus', { target: { value: '2' } });
+    expect(onFocus.mock.calls).toEqual([['2']]);
+  });
+});
+
+describe('Dropdown onBlur', () => {
+  const onBlur = jest.fn();
+
+  const wrapper = mount(
+    <Theme>
+      <Dropdown size="md" items={[{ value: '1', label: '1' }, { value: '2', label: '2' }]} onBlur={onBlur} />
+    </Theme>
+  );
+
+  it('should call onBlur prop', () => {
+    const select = wrapper.find(Select);
+    select.simulate('blur', { target: { value: '2' } });
+    expect(onBlur.mock.calls).toEqual([['2']]);
+  });
+});
+
+describe('Dropdown items properties', () => {
+  it('should disable an item if isDisabled is passed through on the item', () => {
+    const wrapper = mount(
+      <Theme>
+        <Dropdown size="md" items={[{ value: '1', label: '1', isDisabled: true }]} />
+      </Theme>
+    );
+
+    const option = wrapper.find('option');
+    expect(option.props().disabled).toBe(true);
+  });
+
+  it('should hide an item if isHidden is passed through on the item', () => {
+    const wrapper = mount(
+      <Theme>
+        <Dropdown size="md" items={[{ value: '1', label: '1', isHidden: true }]} />
+      </Theme>
+    );
+
+    const option = wrapper.find('option');
+    expect(option.props().hidden).toBe(true);
+  });
+});